Refactor basementCounter to always close the file
To do this I include the break condition in the do loop's condition
instead of a separate if with it's own return statement.  This way the
code after the loop is always run, including the positive return
statement and the file closing function.  Instead I added another if to
check if the end of file was reached, in which case Santa never went
below ground.