Refactor all functions to pass array pointers instead of lots of ints
This has shortened the code and simplified things quite nicely.  It also
made the smallestSide() function completely redundant because, with a
sorted dimension list, the shortest side is made up of the first 2
elements, no extra logic required!  The cubicVolume function cycles
through the array with a for loop and compound multiplies the elements.