Beginnings of Santa navigator code for day 3
There's a struct to contain house info (location & times visited) in a
linked list.  There's a working main loop reading from stdin and ending
on EOF.  The list free-er works in theory but hasn't been checked and
the house add-er isn't done yet.  When it is, the idea is to add a new
house to the list every time Santa moves, unless it's already been
visited, then it will just have its counter bumped.