Refactor count out of house type
It was a sweet idea to include the count as a property of a house, but a
bad one because then it can't be a linked list any more.  Because then
there could be multiple paths to the same house (as in reality) which we
don't want to care about in a list where there is only one path.  Now
house_add returns an integer that can be used to increment a counter
based on whether the house was unique in the list or not and it works as
expected.