Refactor to allow any size of house list arrays
This involved reducing the responsibility of house_count to just
counting.  Now instead of passing it two house lists which it joins and
uniqs, it receives a joined, uniqd list and simply counts it.

To do this I wrote a new function house_join which receives an array of
house lists and joins them by appending each element's first item to the
last item of the previous element and returns a single long house list.

The code in the main function was also improved to use for loops instead
of hard coding how many arrays would be initialised and freed.