Working acoin!  Brings all my helper functions together
For this I added some new helper functions chop and numcat.  Chop chops
off all but the first 5 characters from a string (to make it comparable
with strcmp, if I used strstr it would also match 5 zeroes in the middle
of the string), and numcat uses sptrinf to append an unsigned integer to
the end of a string (your KEY), which lets you give it incrementally
higher key + number combinations until you find the hash whose first 5
digits are 0s.

The heavy lifting is all done in the while loop's condition though,
which makes it unreadably long, maybe that could be extracted out into a
function.