Working solution for part 2 of day 5
That amazing feeling when your code runs correctly on the first try!!!

Both these functions work similarly to the one that looks for doubled
letters in part 1, the difference is here I put an offset of 2, so it'll
find doubles with a character in between and the other one uses another
loop, steps ahead enough characters that the matches won't overlap and
looks for matches of where the outer cursor currently is, in a similar
fashion.  Obviously it doesn't need to look behind itself as well
because those would have matched already.